Floats

+++ src/helpers.h
@@ -1,3 +1,3 @@
-void print_recept(int tempid, char* mednavn, char* besk, int styrke, int frek);
+void print_recept(int tempid, char* mednavn, char* besk, float styrke, int frek);
 int getTerminalWidth();
 void textColor(char col);
